Stub out a monad for doing erl operations.
Erl.Monad includes

- a special ErlError data type for erl errors

- a MonadErl type class that captures basic operations on entity types
  and entities

- an ErlT monad transformer class, built from ErrorT and StateT

- a ErlMonad monad built from ErlT on the Identity monad

- a stubbed out MonadErl instance declaration for ErlT applied to an
  arbitrary monad.

The state carried by ErlT is currently empty, so the operations are
all stubbed out. In the next few commits, I'll add some data to the
state and implement the operations. I'm a bit up in the air about
whether to leave the errors as strings or define a set of specific
errors in advance. The second approach would help with unit tests.