Add unit tests for the entity type method implementations.
The 'tests.hs' program contains the first batch of QuickCheck unit
tests. I've added three separate tests for the effects of adding an
entity type. The 'Name' and 'ErlTState' types have Arbitrary
instances; currently an arbitrary ErlTState is just the result of
adding a generated list of entity types to the empty state.

As of version 1.9.2, Cabal supports building and running unit test
suites. I've added a test suite configuration to the cabal file, and
bumped up the minimum cabal version.

As far as I can tell, the simplest way to integrate QuickCheck unit
tests is to use the 'exitcode-stdio-1.0' test type, where we have to
create a test executable that returns a non-zero exit code if any of
the tests fails.  In order to do this, we need to use the
'quickCheckResult' function for each test, and 'isSuccess' to check
that all the results are successful; the ordinary 'checkCheck'
function only prints test results to standard output, and if we used
it we couldn't return a non-zero exit code on failure.

Cabal also supports a test suite type called 'detailed-1.0', which
allows Cabal to interact with the test suite and do its own detailed
logging. I don't think the lastest version of QuickCheck has
simplified support for this interface yet. I looked quickly and
couldn't find any specific cabal support, so I stuck with
'exitcode-stdio-1.0'.

Unfortunately, we end up having to repeat the dependencies of the main
project in the 'build-depends' part of the test suite
configuration. I'll check the Cabal documentation to see whether there
is a way to avoid this, but I'm not sure that there is.