Discard built-in entity instance names.
Thinking about it a bit, I realized that we'll want to be able
index entities by properties of their associated data, and
have those indexes managed in a MonadErl. This makes an extra
name attribute for instances a bit redundant -- a name could
just be one of those properties we can index by. I've dropped
the 'name' attribute on Erl.Entity.Instance, and also dropped
the Name parameter for MonadErl.createEntity, and completely
removed the method MonadErl.renameEntity.