Prepare to record entity instances close to their entity types.
I've changed the hidden internals of ErlTState a little, replacing
the 'typesById' map with an 'entityTypeRecsBId' map, where the values
are of type EntityTypeRec. An EntityTypeRec holds an entity type
together with the next instance ID for that type, and a map from
instance IDs to instances.

There is one problem with the approach I've taken so far: it doesn't
accommodate 'scoped' entities. I'm thinking of simplifying things
a bit, and not implementing 'scoped' entities by embedding a reference
to a scoping entity in the key of the instance.

We don't currently support a way to create a scoped entity; we also
don't indicate within an entity type that its instances are intended
to be subordinate to entities of another type. So we either need to
enlarge the MonadErl type class with more methods, or we need to
simplify it, implementing 'scoping' using another mechanism.