Add the Erl.EntitySet module.
Erl.EntitySet is just a newtype on Data.IntSet with automatic
conversion between integers and entity IDs. It also helps
limit the interface to just the minimum necessary operations,
and insulates the rest of the code from depending directly on
Data.IntSet.

Ultimately, we'll want to be able to efficiently compute unions,
differences, and intersections of sets of entity IDs. In the
next commit, I'll replace lists of EntityId with EntitySet in
the MonadErl methods.