Split out an Erl.Test.EntityTests module.
I'm partway through implementing binary relations, and I'd like
to add some tests, but I'd rather not create one big monolithic
test program in the tests.hs module. So I've decided to split
out the contents of tests.hs into several modules below Erl.Test.

This is the first, Erl.Test.EntityTests. I realized that in addition
to the Arbitrary instance declaration for EntityId, I should have
a test that shows 'toInt' and 'fromInt' are inverses, so I've added
that.