Add the first part of the Erl.BinRel module, for binary relations.
Although it's not finished, this has taken a little while to write and
I'd like to checkpoint the implementation and tests, so I can switch
to shorter commits for the remaining parts.

Binary relations are represented as pairs of maps from entity IDs to
entity sets, with type

 EntityMap EntitySet

This makes the right image and left image operations easy to
compute. The two maps are redundant in the sense that either map could
be used to convert a relation to a list of pairs, but the idea is to
make it efficient to traverse a relation in either direction.

I've added basic operations and QuickCheck tests for them, but not
'compose', 'union', 'difference', or 'intersection'. I'll implement
these operations in separate subsequent commits.