Implement 'compose' in Erl.BinRel.
I've also added a couple of unit tests in Erl.Test.BinRelTest.
For the tests I found it convenient to expose the BinRel
functions

  rightImage' :: EntityId -> EntitySet
  leftImage'  :: EntityId -> EntitySet

in order to avoid having to either redefine them in the tests
or tolerate a little more long-windedness.

The generated data used with the second test, on prop_compose,
is probably too unconstrained. Two arbitrary BinRels are likely
to have an empty composition, so in the next commit or so I'll
probably write a new generator that makes it more likely the
relations have a non-trivial composition.