Add 'unionWith' to Erl.EntityMap.
It turns out that ordinary union isn't what we want for maps,
at least not as far as implementing BinRel union is concerned.
That should have been obvious from the beginning: ordinary
map unions throw away the values in one of the argument maps
on the common keys. What we want is union with a combining
function, 'unionWith'.