Add 'differenceWith' to Erl.EntityMap.
Similar to the 'union' case, we really need a combining function
for differences on EntityMaps in order to implement 'difference'
for BinRel.

In this particular case, I've chosen to restrict the interface
of Data.IntMap slightly. The 'differenceWith' function in that
module has type

 (a -> b -> Maybe a) -> IntMap a -> IntMap b -> IntMap a

where the second map can contain a different type of values.
In the cases where we'll need it, I don't think we're going
to need the extra flexibility, so 'differenceWith' for
EntityMaps has just a single type parameter:

 (a -> a -> Maybe a) -> EntityMap a -> EntityMap a -> EntityMap a