Implement 'difference' in Erl.BinRel.
The implementation of 'difference' is about as simple as for
'union', except that we have to be careful to remove keys
with empty values in the 'fromLeft' and 'fromRight' maps.
I've moved to the top level the 'nothingIfEmpty' helper function
that was defined in 'compose', and used it in 'difference'.

I've also added a couple of test cases, one particular to
'difference' and the empty relation, and one that uses
membership tests. Like with 'compose', using pairs of
arbitrary relations in the second test case is probably
sub-optimal: two arbitrary BinRels aren't likely to
intersect, so we might not be exercising a good range of
examples. I'll consider fixing that in a later commit,
after implementing intersection.