Add 'intersectionWith', 'foldWithKey', 'filter', 'filterWithKey' to Erl.EntityMap.
Unfortunately, the type of 'intersectionWith' in Data.IntMap precludes
deleting entries during intersection:

  (a -> b -> c) -> IntMap a -> IntMap b -> IntMap c

What we'd really like is for the combining function to have the type

  (a -> b -> Maybe c)

where Nothing means 'delete the corresponding key', like in
'differenceWith'. However, we don't have that option. In order
to implement 'intersection' in BinRel, we'll need to take
the intersection using Erl.EntitySet.intersection as the
combinging function, and then drop the keys whose corresponding
sets are empty.

I've added a slightly restricted form of 'intersectionWith'
for entity maps, with type

  (a -> a -> a) -> EntityMap a -> EntityMap a -> EntityMap a

In order to allow completing the implementation of 'intersection'
in BinRel, I've also added

 filter :: (a -> Bool) -> EntityMap a -> EntityMap a

Just in case they're useful, I've also added 'filterWithKey' and
'foldWithKey'.