Add the Erl module.
This module simple re-exports some stuff defined in other modules,
so that users don't have to worry about the individual modules. I'm
not sure how this is going to work, since Erl.EntitySet and Erl.BinRel
pretty much always need to be imported qualified.

I will be a good place to hang API documentation though, if I ever get
around to writing it.