Add the 'Toys.hs' example.
This just plays with entity instances a little bit, to get
a feel for what kinds of additional functions would be useful.

This example doesn't do a lot -- the data accompanying the
entities is a simple two-field record -- but it looks like
there could be quite a bit of boilerplate as far as attribute
accessors are concerned.

It's too bad that we have to have a single type parameter for
the data; I don't think there is any kind of trick we can
play with existential types, because we need to expose the
type of the data in order to have computations over attribute
values.

I realized that we *do* want to have entity set variables.
This is the real substitute for "entity types". In the same
way that we'll have IDs (names) for binary relations stored
in the ErlTState, we'll also want to have named entity sets.
The named sets don't necessarily imply entity 'types', just
named sets of entities that can be updated. After improving
the test data generation in a few of the BinRel tests,
I think I'll revise the ErlMonad yet again, adding named
entity sets.