Rename ErlTState to 'ErlState' in Erl.Monad.
The next couple of commits will be a minor detour, fixing some little
things that have been bothering me for a bit.

In this first commit, I'm renaming the Erl monad state type,
ErlTState, to 'ErlState'.  After thinking about it a bit, I figured
there is no particular reason to name the monad state after the monad
transformer -- in the same way that there is no reason to name the
error type 'ErlTError' instead of 'ErlError'.

This required trivial compensating changes in Erl/Test/MonadTests.hs
and the Toys.hs example.