Rename ErlMonad to simply 'Erl'.
There seems to be a loose naming convention in the standard monads and
monad transformers:

  - a type class called MonadX
  - a monad called X
  - a monad transformer called XT

For example, in Control.Monad.State, there is

  - the MonadState type class
  - a monad called State
  - a monad transformer called StateT

There is a notable exception in Control.Monad.Error, where there is
'MonadError' and 'ErrorT', but the name 'Error' is used for errors,
and the ordinary monad is simply (Either e) where 'e' is some error
type.

It makes sense then to rename 'ErlMonad' to the simpler, more easily
pronounced 'Erl'.

Again, this requires trivial compensating changes in
Erl/Test/MonadTests.hs and Toys.hs.