Add EntitySetId to Erl.Monad.
This is the first step in adding entity set variables to MonadErl.
This is going to be a bit of a throwback to 'entity types', except
there is no built-in idea that named entity sets necessarily
correspond to 'types' of entities.

EntitySetId isn't currently used, but in the next commit(s) I'll bring
them back, and (I think) require the entity CRUD operations to take
entity set IDs.