Use Data.Set.Set Int instead of Data.IntSet.InSet in Erl.EntitySet.
I'm pretty sure that we're going to have to make EntityId more
than just a newtype wrapper around Int again, so we won't be able to
use Data.IntSet.IntSet for the implementation of EntitySet. I could
be wrong about this, but using IntSet could be a premature
optimization. We can always restore it if it turns out that we
can use IntSet after all.

This is really just a first step. Instead of Data.Set.Set Int,
I'd like to use Data.Set.Set EntityId, but I'll make that
change in the next commit.