Use Data.Set.Set EntityId instead of Data.Set.Set Int in Erl.EntitySet.
This completes the reversion to Data.Set.Set EntityId for EntitySets.
At this point, we no longer use the 'toInt' and 'fromInt' functions
on EntityIds in this module.

Next, I'll have to make similar changes in Erl.EntityMap. Once that's
done, we'll be able to remove the 'toInt' and 'fromInt' functions on
EntityId, and be free to change its representation.