Use Data.Map.Map Int instead of Data.IntMap.IntMap in Erl.EntityMap.
This is the first of three steps to remove Data.IntMap from the
implementation of Erl.EntityMap, and it's similar in spirit to
the second-last commit. Here I've replaced Data.IntMap.IntMap
with Data.Map.Map Int. In the next commit, I'll switch to
using the more direct Data.Map.Map EntityId.

The third step will involve fixing two deprecating warnings for
Data.Map.fold, which is now replaced with Data.Map.foldr.

At this point, everything builds and the unit tests pass.