Use Data.Map.Map EntityId instead of Data.Map.Map Int in Erl.EntityMap.
This is the second step in reworking the implementation of
Erl.EntityMap in order to allow changing EntityId. Now
the 'fromInt' and 'toInt' functions for EntityId aren't used any more
in Erl.EntityMap. I think this is the last use of these functions,
so in another commit or so I should be able to delete them entirely.