Stub out MonadErl methods for creating, destroying, fetching named entity sets.
Like we'll have named binary relations, we should also have named
entity sets. We'll also be able to associate constraints and perhaps
triggered computations with named sets.

This first step just stubs out the methods

 createEntitySet
 destroyEntitySet
 lookupEntitySet

and implements the 'entitySet' utility function that throws an error
if 'lookupEntitySet' results in Nothing.

I'll implement the methods in the next commit, and then consider how
to change the interfaces of the other methods.  All updates involving
entities should be on named entity sets, so we'll either need to add
EntitySetId arguments or embed an EntitySetId in EntityId.