Implement the new entity set methods in Erl.Monad.
This implements the three new entity set methods in Erl.Monad,

 createEntitySet
 destroyEntitySet
 lookupEntitySet

and adds a couple of unit test cases relating the three methods.

I've elected to create functions on ErlState so that the
implementations in the MonadErl instance declaration for ErlT are
trivial. In the next commit, I'll rework the definitions of the other
methods so they are similar. At first, I thought I'd try putting all
the details in the instance declaration, but I find that it makes the
declaration harder to read.

ErlState now has 'nextEntitySetId' and 'entitySets' fields, and
'entitySets' is a map from EntitySetId to EntitySetRec.  EntitySetRec
isn't really necessary at this point, since it's just a wrapper around
an EntitySet, but I'd like to hang things like constraints in the
EntitySetRec. I'm really bringing back everything that was associated
with entity types, just without calling them 'entity types'. :-) But I
think this is better, because 'type' implies some kind of type
checking, but we don't really have an Erl type system.