Create top-level functions from the MonadErl method implementations for ErlT.
This just moves the bulk of the implementations of the methods

 selectEntities
 lookupEntity
 createEntity
 deleteEntity

from the instance declaration for ErlT into some private top-level
functions in Erl.Monad. The method implementations are now all
trivial one-liners, which makes the instance declaration easier
to read. This does make the source longer, but I think that is
outweighed by the better readability.