Rename 'destroyEntitySet' to 'deleteEntitySet' in Erl.Monad.
This renames to 'deleteEntitySet' the method 'destroyEntitySet'
on MonadErl. While I ordinarily think of 'destroy' as the opposite
of 'create', I've already got 'deleteEntity', and the acronym
CRUD stands for "CReate, Update, Delete" IIRC. So we might as
well be consistent.

This required some minor compensating changes in the tests.