Add 'entitySetIds' to ErlMonad, and change the type of 'createEntity'.
This commit is a little more involved than the last few, particularly
in the unit tests. I set out to change the type of 'createEntity' so
that you have to give an entity set ID.  The type was

  d -> m EntityId

but is now

  EntitySetId -> d -> m EntityId

where it is an error if the given EntitySetId isn't known.

I ended have having to make the commit larger in order to accomodate
the change in the unit tests.

The tests for 'createEntitySet' needed to be changed to take into
account the new EntitySetId parameter, and needed to take into account
the situations where the EntitySetId is not known, and where it
is. This meant that an EntitySetId needed to be generated alongside
the ErlState and the associated value. In order to make sure that the
EntitySetId was usually drawn from the ErlState, I needed another
MonadErl method,

  entitySetIds :: m [EntitySetId]

and I needed to add a generator 'withEntitySetIdAndVal' for generating
test data for entity creation.

I also had to export a function 'entitySetId' for creating an
EntitySetId from an integer so that I could write an Arbitrary
instance for EntitySetId. For good measure, I also added 'binRelId'
for BinRelIds, figuring that I'll eventually need that when I add
methods for dealing with binary relationships.

While doing all this, I ended up adding some infrastructural functions
to the MonadTests, including

 checkAll
 checkFails

and I decided to use 'checkAll' to combine

 prop_createEntity1 and
 prop_createEntity2

into a single test property

 prop_createEntity

and similarly combine

 prop_deleteEntity1
 prop_deleteEntity2

into a single property

 prop_deleteEntity

Originally I split them in order to make it easier to pinpoint the
problem if a test failed, but I decided to instead create a single
'contract' property for each method. We'll see if that was a good idea
or not :-)

The Erl.Monad module also gets the new utility function

 modify'' :: (MonadError e m, MonadState s m) =>
             (s -> Either s (a,s)) -> m a

for lifting into the monad a pure function on the state that either
fails with an error or returns a result and a new state.

I probably should have left combining the test properties for
'createEntity' and 'deleteEntity' for a separate commit, but I got a
little carried away ;-)

At this point, the internals of ErlState aren't quite right. Entities
are created associated with an EntitySetId and so presumably belong to
the indicated entity set, but they aren't currently stored with their
entity sets. After adding an entity to an entity set 'S', the set 'S'
doesn't in fact contain the entity. I decided to fix that when I
modify the type of 'selectEntities' in the next commit. When I do
that, I'll probably move 'selectEntities' out of MonadErl; a more
general form for 'selectEntities' would have the type:

 (MonadErl d m) => EntitySet -> (d -> Bool) -> m EntitySet

where the required primitive operation in MonadErl is

 getEntityAttributes :: EntityId -> m d

That the current tests still pass indicates that I'm missing a test on
the side-effects of 'createEntity'.