Add 'lookupEntityAttributes' to MonadErl.
In order to change the signature of 'selectEntities' and move it to
the top level, we need a way to get at entity attributes by entity
ID. Strictly speaking, 'lookupEntity' already gives us a way to do
that. However, I realized that we could completely hide 'Entity' and
simply provide 'lookupEntityAttributes' instead.

I've added the new method, and will eliminate 'lookupEntity' in the
next commit. I've also added

 getEntityAttributes

which is like 'lookupEntityAttributes' but throws an error if the
entity ID is unknown.