Remove 'lookupEntity' from MonadErl.
We should probably make 'Entity' completely internal in Erl.Monad;
it only serves to carry the attributes associated with an entity ID,
and generally users with be working with EntitySets. Now that we
have 'lookupEntityAttributes', there is no reason to expose 'Entity'
through 'lookupEntity', so I've removed

 lookupEntity

and the associated top-level function

 getEntity

This required some very minor changes in the unit tests.