Discard the 'Entity' type.
At first, I was going to move Entity from Erl.Entity to Erl.Monad, but
I realized that ErlState needs only a map from EntityIds to the
associated attributes, so 'Entity' can be removed entirely.

This required some minor changes in Erl, and some changes in the
'Toys' example, which had in fact become wrong in the last couple of
commits when I changed the type of 'createEntity'.  I've updated
'Toys' so it creates an entity set 'toys' and places all the toys in
that set.