Remove Entity.toInt.
We still need Entity.fromInt to implement the 'Arbitrary' instance for
EntityId, but we can get rid of 'toInt'. We may want to change
'EntityId' so it carries an EntitySetId, in which case either the
'Arbitrary' instance will need to change, or we'll need to resort to
more constrained generators in the tests.  I'd like to think about
that a little bit more before doing it though.