Add some blather on named relations and attributes.
I think that I want to separate the existence of entities from the
existence of entity sets after all, but associate extra data with
membership in an entity set or a binary relation. This gives EntitySet
and BinRel a more uniform treatment as unary relations and binary
relations, and I *think* that's good thing.

This will complicate a 'destroyEntity' operation, since we'll have to
track which sets and binary relations an entity participates in so we
can fail the operation or do a 'cascading delete'. But I think that's
acceptable.