Simplify 'createEntity' and add 'hasEntity'.
I'm switching to stand-alone entities, as described in
ideas/named-relations-and-attributes.txt. Instead of creating
an entity within an existing entity set and with extra data,
we allocate the entity ID separately. Eventually, we'll associate
the extra data with an element of a particular EntitySet,
by having a MonadErl method

 addEntity :: EntitySetId -> EntityId -> d -> m ()

and a method

 removeEntity :: EntitySetId -> EntityId -> m ()

(the signatures may change). There will be similar methods
for dealing with BinRels, except they'll take two EntityIds.

At this point, there is some unused stuff in the unit tests,
but I'll keep it around for a commit or two to see if I need
some of it. The Toys example compiles, but the functions for
getting name and age don't work, since 'createEntity' no
longer adds data.

In the next commit, I'll add the methods for adding and removing
entities on EntitySets, and I should be able to fix up the
Toys example, and slightly rework the internals of ErlState.