Fix 'lookupEntityAttributes', implement 'addEntity' and 'removeEntity'.
This commit moves the entity attribute storage into the internal
EntitySetRec type, and changes the type of 'lookupEntityAttributes'
from

 EntityId -> m (Maybe d)

to

 EntityId -> EntitySetId -> m (Maybe d)

That is, an entity has associated data only within the context of a
named entity set.

I've also added implementations of 'addEntity' and 'removeEntity',
which add and remove an entity and its associated attributes to a
named entity set. These methods will fail if either the entity set
doesn't exist or the entity doesn't exist; however, they are otherwise
idempotent. That is

  addEntity eid x esid >>
  addEntity eid x esid

does not fail and has the same effect as just

  addEntity eid x esid

and similarly

  removeEntity eid esid >>
  removeEntity eid esid

does not fail and has the same effect as

  removeEntity eid esid.

There are a few issues that need to be fixed in subsequent commits:

  - destroyEntity should fail if the entity is a member of any named
    entity sets; currently, it succeeds, but the entity is not really
    destroyed if it still appears with associated data in entity sets.

  - selectEntities now has a nonsensical type. It's type is

      (d -> Bool) -> m EntitySet

    but entities have data only by virtue of being members of named
    entity sets, and different sets may of course have different
    data. For now I've simply removed the implementation of the
    method. A more general form of the function would have the form

      EntitySet -> (EntityId -> m Bool) -> m EntitySet

    where we filter the given entity set according to a predicate that
    returns its result in the monad and so can perform any query
    supported by the monad.

    This in fact allows 'select' to be a top-level function outside
    MonadErl. Ideally, we'd have a second read-only monad in which to
    do the query, since in this form the query could modify the state,
    which probably isn't what we want.

  - updateEntity similarly has a nonsensical type. It should modify
    the attributes associated with a particular named entity set.

However, at this point, the Toys.hs example works again. The unit
tests all pass without modification, since there are no tests that
exercise lookupEntityAttributes, addEntity, or removeEntity. I'll
probably write some tests for those in the next few commits, but I
might do a little clean-up first, or fix the definitions of
'selectEntities' and 'updateEntity'.