Add a top-level 'selectEntities', and fix {add,remove}Entity.
I've moved 'selectEntities' out of MonadErl and given it the type

 (MonadErl d m) => (EntityId -> m Bool) -> EntitySet -> m EntitySet

I (re)discovered in the process that the unit tests were using
'selectEntities' to get a list of all the entities, so I added to
MonadErl the method

 entityIds :: m [EntityId]

I tried out a little manual testing with 'selectEntities' in the
'Toys' example, running

 evalToys' (getEntitySet toys >>=
            selectEntities ((45 ==) . getAge)) initialToys

to confirm that it picked out 'gordon'. At that point I discovered
that 'addEntity' and 'removeEntity' have a bug: they don't update the
'entities' field of EntitySetRec!

I've fixed that and now the manual tests on 'selectEntities' seem to
do what they should.

This approach to 'selectEntities' allows the predicate to modify the
state. I realized that I could have prevented this by giving
'selectEntities' the type

  (MonadErl d m) =>
    (EntityState d -> EntityId -> Either ErlError Bool) ->
    EntitySet -> m EntitySet

where the predicate takes the monad state and can return an
error. It's possible to implement such a predicate using the Erl monad
with the 'evalErl' function, and this form would prevent the predicate
from updating the state in the enclosing monad. However, I figured
that this inconveniences the user of the librxary simply to prevent a
programming style that is not strictly speaking wrong, but just
ill-advised.  And chances are the user wouldn't modify the state
anyway.  It's certainly easier on the user to put the predicate's
result in the monad and just say in the API docs that it's probably a
bad idea to update the state in the predicate, since it puts
side-effects in an unexpected place.