Make 'updateEntity' a top-level function and correct its type.
I was going to correct the type of 'updateEntity' and then
realized that with 'addEntity' and 'getEntityAttributes' this
doesn't need to be a method on MonadErl. Instead, it can
be a top-level function. I decided to make the type

 (MonadErl d m) =>
   EntityId -> (d -> m d) -> EntitySetId -> m ()

so that it was possible for the update operation to make
queries in the monad.

The implementation fails if either the entity set or
the entity ID is unknown; I think that this is reasonable,
but need to try this out with some examples before I can
tell for sure.