Rename 'lookupEntityAttributes' to 'lookupEntity', and corresponding changes.
I figured that 'lookupEntityAttributes' is kind of long-winded.  Since
the only thing we can look up that is associated with an entity is the
extra data, I figured I'd take a risk and just cal the method
'lookupEntity'. This is also consistent with 'addEntity' and
'updateEntity'.

I've also changed the related functions like 'getEntityAttributes' and
the internal helper stuff.

At this point, I think I've finally implemented (but not properly
tested) all the basic methods that deal with entities and named entity
sets. I'll either work on the unit tests or start adding the methods
that deal with named binary relationships and their associated data.

I had to make a minor compensating change in the 'Toys' example.