Cleanup: factor out 'noSuchSet' and 'noSuchEntity' utility functions.
I throw "no such set" and "no such entity" errors in a couple of
places, so there might as well be utility functions for this.

This will also make it easier to adapt if ErlError changes to have
specific variants for common errors such as these. In fact, I think
I'll do that in a later commit, in order to make it easier to write
more specific unit tests.