Don't import Erl.Entity qualified in Erl.Monad, Erl.Test.MonadTests.
We really need only the EntityId type and its eponymous constructor,
and the name doesn't conflict with anything. It's also annoying
to write 'E.EntityId' instead of 'EntityId' all over the place.

This changes many lines in Erl.Monad, but other than the change
in import, the changes simply replace 'E.EntityId' with 'EntityId'.