Use the same warnings when building 'tests' as when building the library.
I noticed in the last commit that there were no unused import
warnings in MonadTests when there should have been -- in particular,
I had the imports

 import Erl.Entity (EntityId)
 import qualified Erl.Entity as E

when only the first was necessary, but there was no warning when
the tests were rebuilt.

Looking at the cabal file, I realized that the section for the
test suite didn't have any 'ghc-options' set. After adding the
same GHC warning options as in the main library section, I
found that several unused imports had crept into the tests.

This updates the erl.cabal file and removes the unnecessary
imports.