Implement {create,delete,lookup}BinRel, and 'binRelIds'.
This implements the MonadErl methods

 createBinRel
 deleteBinRel
 lookupBinRel
 binRelIds

These methods are essentially copies of the implementations for the
corresponding entity set methods. That's not very satisfying, but
we're not talking about a lot of code, so it's not worth worrying
about for the moment.

I'll add unit test cases in a separate commit.