Rename 'checkAll' to 'allOf', 'checkNoneOf' to 'noneOf'.
In the last commit, I added 'checkNoneOf' to Erl.Test.MonadTests.  But
the name 'checkNoneOf' is a little confusing, since it sounds like it
should not bother executing any of its conditions.

I considered renaming these functions 'assertAll(Of)' and
'assertNoneOf', but 'assert' suggests that there is an immediate
failure at the first 'False' (or 'True'), which is also a little
misleading. So I settled instead on the names

 allOf
 noneOf

which suggests that the result is a boolean, as it indeed is.