Record the participation of entities in binary relations.
I've added a field

  inBinRels

to EntityRec in Erl.Monad, in which to record the named binary
relations that an entity is participating in. The 'addPair' and
'removePair' functions update this field for each entity ID in the
pair.

In order to support this change, I added a function 'participatesIn'
to Erl.BinRel, which tests whether an entity ID appears in either the
left or right roles of a binary relation.

This change is a bit dissatisfying, because in 'removePair' the
'entities' map in the ErlState gets updated even when there is no
change in participation for either of the entity IDs. It would be nice
to avoid updating the map if it's not necessary, but the code is
already a little messier than I'd like. Perhaps the extra work is so
minor that this won't be a big performance problem, and I don't need
to worry about it.