Add unit tests for 'addPair' and 'removePair'.
I've added two new test properties,

  prop_addPair
  prop_removePair

that exercise the contracts of the 'addPair' and 'removePair' methods
on Erl. These are similar in spirit to the 'prop_addEntity' and
'prop_removeEntity' functions, except we have to account for either of
the entity IDs being bogus.

As with several of the other tests, the test data generation might not
be ideal here -- we might not be getting a lot of test cases where the
BinRelId and both EntityIds exist -- but this is a good start.