Move some code in Erl.Monad, Erl.Test.MonadTests.
It makes sense to put the MonadErl methods

  createEntity
  deleteEntity
  hasEntity
  entityIds

first, since allocating entity IDs is a prerequisite for adding
entities to entity sets or pairs to binary relations. I've similarly
moved the definitions in the instance declaration for ErlT, and the
supporting code, and the test properties

  prop_createEntity
  prop_deleteEntity

in MonadTests.

There are absolutely no functional changes in this commit.