Balk when deleting an entity that's in a set or a binary relation.
We're tracking when an entity is a member of a named set or
participates in a named binary relation. However, we don't throw an
error in 'deleteEntity' when the entity is used by a named set or a
named binary relation. This can lead to inconsistencies, since the
methods

  addEntity
  removeEntity
  addPair
  removePair

all balk if an entity ID isn't known.

I've added a new special error case

  EntityInUse EntityId [EntitySetId] [BinRelId]

and changed 'deleteEntity' so that it throws this error if the entity
is in a set or participates in a binary relation.  I've also changed
it so it throws NoSuchEntity if the entity doesn't exist; I'm not sure
if that's a good idea since it makes 'deleteEntity' non-idempotent,
but I'll leave it that way for now.

In order to fix up the unit tests, I realized that the tests had to be
able to predict whether 'deleteEntity' would fail with an
'EntityInUse'. In order to allow this, I added two new methods to
MonadErl,

  memberOfEntitySets
  memberOfBinRels

I'm not sure I like the names, but I'll consider changing them later.

As part of this commit, I also decided to enhance the Arbitrary
instance for (ErlState d) so that it

 - adds some number of binary relations

 - adds entities to some of the entity sets

I'll add pairs to some of the binary relations too in the next
commit. I've not done that here because the tests have gotten
noticeably slower, and I think the problem is that I need to limit the
number of entity sets instead of creating arbitrarily many of
them. I'll see if that helps before moving on to other changes.

There is still an inconsistency in the way we track an entity's
participation in sets and binary relations: when we destroy a set or a
binary relation, we don't remove the EntitySetId or BinRelId from the
'inSets' and 'inBinRels' sets associated with each participating
entity. That's probably a large enough change that it needs to be made
in a separate commit.