Limit the number of entity sets, binary relations, and set contents in Arbitrary ErlStates.
I've added a utility

 atMost :: Int -> Gen a -> Gen a

that limits the size associated with a generator to a given limit. For
size-sensitive generators like 'listOf', or 'arbitrary' list
generators, this limits the size of the lists.

I've used this in the Arbitrary instance for (ErlState d) to limit the
number of sets and binary relations to at most ten each, and the
number of elements of a set to at most twenty.

This seems to noticeably speed up the tests.

I'll have to try using HPC to check that the tests are exercising all
the code I think they are; there are a few things I'd like to do first
though.