Generate IDs to delete for 'delete{Entity,EntitySet,BinRel}'.
Instead of always using arbitrary IDs, we should ensure that the IDs
are usually drawn from the ErlState. I've added some new generator
functions

 withEntityId
 withEntitySetId
 withBinRelId

that generate a (ErlState d) together with an ID that is drawn from
those in the ErlState, provided there are any, and an arbitrary ID
otherwise. I've changed the signatures of

 prop_deleteEntity
 prop_deleteEntitySet
 prop_deleteBinRel

so they take the IDs to delete in an argument tuple, and I've changed
the tests to use 'forAll' with the new generators. Although I never
measured it, I suspect that much of the time the 'prop_deleteEntity
'test was deleting an entity ID that wasn't in the ErlState to begin
with.  The 'prop_deleteEntitySet' and 'prop_deleteBinRel' tests first
created an ID and then deleted it, so they never exercised the case
where the ID was unknown.

At the moment, 'deleteEntitySet' and 'deleteBinRel' do not balk when
the ID doesn't exist. On the one hand, this makes these operations
idempotent; on the other hand, it's inconsistent with the
'deleteEntity'. I'll decide in the next little while whether I want to
preserve idempotency, in which case I'll need to change
'deleteEntity', or whether I want to emphasize 'sanity' checking.

In the process of adding the new generator functions, I decided to
rename the 'sampleX' functions to 'someX'. So now

 sampleEntityIds
 sampleEntitySetIds
 sampleBinRelIds

are called

 someEntityIds
 someEnitySetIds
 someBinRelIds

I've added corresponding generator functions

 anEntityId
 anEntitySetId
 aBinRelId

that generate IDs from an ErlState, if there are any, and arbitrary
IDs otherwise.