Add pairs to binary relations in the Arbitrary instance for (ErlState d).
A couple of commits back I added random lists of existing entities to
the generated entity sets. In this commit, I've done the analogous
thing for the generated binary relations. In the process, I've
extracted another helper function

 subLists :: [a] -> Gen [a]

from 'someOf'. This is similar in spirit to

 listOf . elements

except that it handles the case where the input list is empty;
'elements' invokes 'error' if its input is empty. Hoogle can't find an
equivalent in QuickCheck, but it's no big deal to add it.