Make 'deleteEntity' idempotent: don't balk at non-existent entity IDs.
I was going to change 'deleteEntitySet' and 'deleteBinRel' so they
balked at non-existent entity set IDs and binary relation IDs, but
then I realized this would make them non-idempotent. That is

 deleteEntitySet eid >> deleteEntitySet eid

would fail with a 'no such entity ID' error in the second
'deleteEntitySet'. I think it's better for this operation to be
idempotent, since it's more convenient, so I've decided to leave
'deleteEntitySet' and 'deleteBinRel' alone and change 'deleteEntity'
so it does nothing on a non-existent ID.

This does require a compensating change in the unit tests, a minor
simplification.