Remove entity ID -> entity set ID associations when destroying an entity set.
When we destroy an entity set, we have to remove the entity set ID
from the 'inSets' field for each entity ID in the set. Otherwise, it's
impossible to destroy the entity. I've modified 'esDeleteEntitySet' so
it performs this extra work for each of the members of the set.

I've updated the unit test property 'prop_deleteEntitySet' to check
for this. I've also verified that without the change in
'esDeleteEntitySet', the tests reliably fail, and with it they
succeed.